first cut on most of the totals functionality.  most of this is to read the various totals.  only things really functional at this time are:  - able to log into a GenPOS terminal and log out  - able to retrieve the cluster financial totals
The other totals have not been tested yet and most of them are not ready
for testing. Coupons, Cashiers, Departments all need to have some way of
pulling the list of existing items and to then iterate over the list to
pull the totals for each type of data.

We also need to review the totals reset for both a current total and for
the End of Day.