changes to Cashier Edit dialog and UI View
updated the Cashier Edit dialog to provide controls to allow editing of
the Cashier data. we introduce a tab control to provide the check boxes
for the various on/off settings for a Cashier record.

still have work to do in this area. mainly just to insert the tab
control and the tab windows.