added in functionality to pull EJ report and save to file
New functionality added to provide for a worker thread to handle
activities that may take more than a few seconds. this allows the UI to
remain responsive.

Candidate activities to use the worker thread include long time
transfers such as reading the Electronic Journal (EJ) and saving the
data to a file.

To provide a test for the worker thread, we have implemented the reading
of the EJ and writing the data to a text file. The text file can be
opened with NotePad or other text editor for reading or printing.