prepare for totals output of text, JSON, and CSV
created a function that accepts a union for the various total types
(long, long long, TOTAL, etc.) and lookup tables to create strings for
various totals report output formats. we are looking to provide standard
text, JSON, and CSV output of total data retrieved from a terminal.

currently the output is only going to the Output window of the debugger
using TRACE() calls so that we can work out the details.