added new DialogMnemonic class for display of mnemonics
We have a number of different types of mnemonics to display and edit. We
introduce a class to display those various kinds of mnemonics. This is a
dialog that contains a CListCtrl at the moment to be used for displaying
a list of mnemonics. Still have work to do to use this new class for the
various mnemonics. Also need to take a look at how to do an edit of a
mnemonic.