added first cut of reading several mnemonics
added read of Transaction and Lead Thru mnemonics to Retrieve command
which retrieves the settings. the mnemonic retrieve is working and we
can see the data in the debugger.

still need to hook this into the mnemonic dialog and allow edits and
changes to the mnemonic data.

still need to test the writing or updating of the terminal.