P20 Transaction and P21 Lead thru mnemonics added
Added to the resource.h and the .rc file the defines and necessary
strings to provide the prompts for editing transaction mnemonics (P20)
and lead thru mnemonics (P21). made changes to class CParamMnemonic and
those classes derived from it to provide a list of mnemonics of the type
of the class derived from CParamMnemonic using polymorphism.

we have the iteration code in place and the design of the mnemonic edit.
the mnemonics seem to be put into the CListCttrl in reverse order so
need to figure that out.

This model can be used for the remaining mnemonics.