prototype PLU edit dialog changes placing new Tab Control
worked on appearance of the PLU edit dialog which displays the current
settings for an individual PLU and allows the settings to be changed.

Added new tab control along with two pages of the status bit settings
similar to the Cashier Edit dialog box.

Moved CTabCtrolDialog which was in CDialogCashierEdit source files (.h
and .cpp) into a new file that contains dialog utility classes.

Generalized the use of the CTabCtrlDialog in the CDialogCashierEdit
class so that similar mechanics are available for CDialogEditPlu.