improve LanLogin dialog to capture login and memo
There are two different dialogs for the actual login information such as
host name, IP address, whether to use DHCP or not, password, etc.

CDialogLan is the standard information that is saved in the archive or
file describing the settings of a terminal and cluster. CDialogLanLogin
is the dialog presented to actually perform a login to a terminal.

We have made changes so that these two different dialogs are now
synchronized through the document class data.

So you can use the CDialogLan dialog to setup the information about the
cluster and then when using the CDialogLanLogin dialog to log in, the
login data will be propagated correctly.

We have also made the first change to the Serialization of the document
to save and retrieve the CDialogLan related data.