Make Lan Connection Details a button on main screen
Make the LAN Connection Details dialog be displayed by a button which is
added to the main application window. This button is the same as the Lan
Connection Details dialog trigger as the Lan Connection ... menu item
from the View menu. Pull this button out to make it easy and quick to
find.