create LAN Connection class to manage LAN connection data better
In order to compartmentalize and encapsulate the LAN connection data we
have created a class to hold all of the data pertinent to a LAN
Connection. We have extended the class a bit more by adding operator >>
and operator << functions for CArchive and CLanConnectionData to perform
the serialization for the object.