replace magic constants with defined constants for mnemonic length
Replace the magic constants used for mnemonic length of number of
characters with the actual defined constant needed for the type of
mnemonic.

We are seeing compiler warnings indicating a need to convert wcsncpy()
to the secure version, wcsncpy_s() however the behavior of the secure
version is not exactly what we want. We really want to copy at most the
specified number of characters and not have to deal with the exception
that may be thrown.