added Tender Key Edit dialog
Blocked out the Tender Key Edit dialog. This dialog will present the
settings to be changed for a particular Tender Key. We are using a tab
control to present the various MDC Address pages for the Tender Key in
order to allow all settings to be consolidated into a single dialog.

We will be using this same approach for other keys such as Total Key to
provide a single place for editing the behavior of a particular key.