correct PLU retrieval of duplicated PLUs at end with Table Service database
Working with the table service database with GenPOS Rel 2.2.1 we are
seeing that the last set of PLUs, five PLUs which is exactly the number
retrieved each time the function SerOpPluAllRead() is called, are being
repeated. In other words we call the function to retrieve the last five
PLUs which returns an error code of zero and we add those five PLUs to
the list. We then call the function again which returns a value of zero
however no new PLUs were returned and the existing values remain so we
add the same PLUs to the list of PLUs. Then when the function is called
again, we get a negative value as the return code.

We have introduced a new method to the CParamPLU class to clear the PLU
numbers so that determining if a PLU was returned or not is a check for
the PLU number to be non-zero.

The behavior may be a defect in GenPOS itself and should be looked into.
If a call is made and there are no more PLUs because the last five were
returned in the previous call then we probably should return a negative
value indicating we are done.