changes to Serialization to include mnemonics, management info
Starts some of the changes to the Archive Serialization functionality.
First of all we are now using signatures or special identifying bit
patterns in places so as to help check the Archive file integrity.

We are also implementing a version number value which is the first item
in an Archive to indicate the application version which wrote the
Archive.

We have added Serialization of the Transaction and Leadthru mnemonics.
to the Archive Serialization functionality for the document.