change solution to MDI without OLE and without separate frames, add SQLite3
The previous MFC usage was of using Multiple Document Interface (MDI)
with separate frames for each of the documents. We also had OLE support
for embedding objects. These solution settings resulted in an
application that was not intuitive when opening up existing databases
and the OLE plumbing was just in the way.

We have changed this to use a single frame with multiple documents
within the frame.

We have also added the SQLite3 embedded database engine as a project.
The intent is to move forward with using SQLite3 as the persistent data
store technology for provisioning data. We will maintain the actual data
as blob data while using a key created by some unique value of the data
blob (cashier id number or PLU number for instance).