working on using SQLite3 database for storing provisioning data
First step in using SQLite3 database engine to store the terminal
provisioning data.

We have a prototype of pulling some of the provisioning data from a
terminal. We are currently building an SQLite3 database file containing
the following data:
- PLU database
- Cashier/ETK data
- Transaction Mnemonics
- LeadThru Mnemonics

The current status is being able to pull the data from the terminal. We
do not yet allow the editing and viewing of the data in the SQLite3
database. We do not yet allow the pushing of the data in the SQLite3
database to the terminal.