created new animated progress control derived from CStatic to show communicating
Since communication over the LAN to a terminal when doing actions that
take some appreciable time such as downloading from the terminal a copy
of the provisioning data or uploading to the terminal new provisioning
data, we introduce a new animated progress control.

This first version uses a CStatic to show numbers counting. We need to
do additional improvements to provide a better idea as to status such
as:
- doing preliminary actions such as connecting to the LAN
. may need to show status such as in progress, success, or failed
- doing the actual requested actions such as upload or download
- doing ending actions such as disconnecting from the LAN

There is still a bit of work to do with this control.