added new DispatchToAllViews() method to send an ON_MESSAGE() message to all views
Adding a new method, DispatchToAllViews(), so that we can send a message
to all of the views for a document.

We use this method to inform all views of document actions which are
outside the normal MFC messages. We will be using this to inform all
views when there is some kind of an action that will take some time.
Examples would be provisioning download or upload from a terminal. The
idea is to let the views know it should display some kind of user
visible notification.