changes to the view to handle a new WM_USER type message with CProgressWnd
We have implemented a first cut on the use of the new CProgressWnd
control so that when doing time consuming actions in a separate thread,
we can show a progress indicator to the user.

We will need to do further work to extend the functionality of the
CProgressWnd control to provide more flexibility. Currently we just
toggle it off and on however we probably should include the ability to
specify various kinds of appearance changes.