implemented use of CProgressWnd progress indicator with LAN activity
We now have a method for sending a message to the LAN thread specifying
an action to perform as well as a function call back to start and stop
the CProgressWnd progress indicator.

The CProgressWnd progress indicator is currently primitive in that it
increments a count every half second and then displays the new count as
a text string in a visible CStatic control. However this works well
enough for now.