first phase cleanup and refactor of existing functionality
We have done a bit of cleanup with the various classes and removing the
'virtual' keyword where it was specified but should not have been in
derived classes which should be the final.

We have also updated comments in various places.

We are now working on refactoring the source to make the interfaces to
the various types of data fairly generic and regular.

We are treating all data as a collection of individual data items that
compose a map of some key to a particular data item. In most cases the
key will actually be a binary number much like what is used in GenPOS as
an address. For some data collections the key will be a string of
characters for instance a PLU number as a key.

The key is used to access the data blob which is the actual GenPOS data
struct of the data as it is retrieved from a GenPOS terminal.

We are mirroring this same list in the SQLite database file in which
each collection is a separate table within the database. The database
file itself currently contains a single database however this may change
in the future.

We still have some work to do to better integrate use of SQLite database
engine for persistent storage of data.

Much of the work in this current update deals with transaction and lead
thru mnemonics. We have been concentrating on refactoring the previous
code where the list was part of the CParam derived class in CLister
derived classes. The idea of these changes is to make the mnemonics
similar to PLUs, cashiers, and other data. Most of the mnemonic data is
stored in the memory resident database in GenPOS hence is really stored
as arrays of fixed size whose elements are character arrays of a fixed
length.