refactoring with lessons learned on mnemonics and databases
We are now working on the changes to provide the necessary SQLite
routines for the following kinds of actions:
- retrieve data from GenPOS and store in memory list
- retrieve data from GenPOS and store into database
- retrieve data from database and store in memory list

The idea is to be able to populate the memory list from either GenPOS or
from a SQLite database file.

Once the data is into the memory list, we can then display and allow for
edits from the source which would be either GenPOS directly or the
SQLite database file.