replaced TCHAR with wchart_t and _T with L. replaced a few magic constants with defines
This was just some simple cleanup and eliminate some compiler warnings.

Instead of TCHAR we are using C++ standard wchar_t for text character
arrays. Also replaced use of the +T() macro with leading L on text
string constants.

A couple of places we were using magic constants and we have replaced
those with the correct defined constant.

Also fixed a couple of warnings about fcopy() and replaced it with
fcopy_s() instead.

We are still using wcsncpy() instead of the secure version as we really
need the behavior of copying up to the max and not putting a trailing
zero terminator if there is no more room. We really do not need any
exceptions being thrown if the exact maximum of characters are copied.