clean up the file open for the database file when uploading/downloading
We are now using the database name specified in the Login Settings for
the name of the SQLite database file.

If a name has not been specified then we will ask for a file name.

We are also doing a SetCurrentDirectory() call to set the default folder
so that all files will be stored in the same folder unless the user puts
them somewhere else.

We are currently using using the message handler OnTerminalEJretrieve()
which is triggered by doing an EJ Retrieve from the menu. We still have
some work in order to finalize the command set and the look and feel.