change the main window designer to better group the various controls
We have re-arranged some of the buttons on the main window. Previously
the buttons were just arranged in a row across the top of the window.

Now we have the Login Settings button on the left above the display of
the current directory, database name, etc as before.

However the Log In and Log Out buttons were moved to where the Retrieve
button was located and the Retrieve buttons dropped down a row. This
change locates the Log In and Log Out buttons along with the Retrieve
button in a 2 x 2 grid next to each other. This grouping makes it
obvious that this group of buttons are used together.