first draft integrating existing database into application
Modified the Retrieve menu so that now we can either retrieve over the
LAN from a terminal or retrieve from an existing database file.

Still need work on figuring out use of database file and allowing for
changes to a database file by doing a retrieve from a terminal over the
LAN or by doing edits to an existing database previously created.

With these new changes we can now display the list of Transaction
Mnemonics and Lead Thru Mnemonics that are stored in an existing
database file.