changes from LAN centric to database centric
The first versions of this application were from a perspective of a tool
that pulled data from a terminal and displayed that data. The database
was a way of just storing the data.

The application was LAN centric rather than database centric since the
emphasis was on directly modifying the terminal data with the database
as being just storage.

We are now starting a change to make the application database centric.
The idea is to pull data from the terminal into a database and then
view/edit the data in the database.